+++ src/index.ts
@@ -4,6 +4,8 @@ export * from "./preprocessing/StandardScaler";
 export * from "./linear_model/LinearRegression";
 export * from "./linear_model/LogisticRegression";
 export * from "./neighbors/KNeighborsClassifier";
+export * from "./tree/DecisionTreeClassifier";
+export * from "./ensemble/RandomForestClassifier";
 export * from "./model_selection/trainTestSplit";
 export * from "./metrics/regression";
 export * from "./metrics/classification";